+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 1.2.0 (...)
+- add `setFile($fileName)` method to `Response` class which sets the
+  `X-SendFile` header (for mod_xsendfile)
+
 ## 1.1.3 (2015-11-01)
 - undo changes of 1.1.2, needs furter investigation to be sure we are
   taking the right approach.